vortex/backend-py-personal/requirements.txt
fastapi
uvicorn
google-adk>=0.7.0
python-multipart
requests
pydantic
python-dotenv